refactor
reorganize files and components
