+++ src/logic/safe/store/actions/fetchTransactionDetails.ts
@@ -2,11 +2,11 @@ import axios, { AxiosResponse } from 'axios'
 import { createAction } from 'redux-actions'
 
 import { getTxDetailsUrl } from 'src/config'
-import { Dispatch } from 'src/logic/safe/store/actions/types.d'
-import { ExpandedTxDetails, Transaction } from 'src/logic/safe/store/models/types/gateway.d'
+import { Dispatch } from 'src/logic/safe/store/actions/types'
+import { ExpandedTxDetails, Transaction } from 'src/logic/safe/store/models/types/gateway'
 import { TransactionDetailsPayload } from 'src/logic/safe/store/reducer/gatewayTransactions'
 import { safeParamAddressFromStateSelector } from 'src/logic/safe/store/selectors'
-import { getTransactionDetails } from 'src/routes/safe/components/GatewayTransactions/Expanded/selectors/getTransactionDetails'
+import { getTransactionDetails } from 'src/logic/safe/store/selectors/getTransactionDetails'
 import { AppReduxState } from 'src/store'
 
 export const UPDATE_TRANSACTION_DETAILS = 'UPDATE_TRANSACTION_DETAILS'
