+++ src/routes/safe/components/GatewayTransactions/hooks/useQueueTransactions.ts
@@ -1,0 +1,37 @@
+import { useEffect, useState } from 'react'
+import { useSelector } from 'react-redux'
+
+import { TransactionDetails } from 'src/logic/safe/store/models/types/gateway.d'
+import { nextTransactions, queuedTransactions } from 'src/logic/safe/store/selectors/gatewayTransactions'
+
+type QueueTransactionsInfo = {
+  next: TransactionDetails
+  queue: TransactionDetails
+}
+
+export const useQueueTransactions = (): QueueTransactionsInfo => {
+  const nextTxs = useSelector(nextTransactions)
+  const queuedTxs = useSelector(queuedTransactions)
+  const [txsCount, setTxsCount] = useState({ next: 0, queued: 0 })
+
+  useEffect(() => {
+    const next = nextTxs
+      ? Object.entries(nextTxs).reduce((acc, [, transactions]) => (acc += transactions.length), 0)
+      : 0
+    const queued = queuedTxs
+      ? Object.entries(queuedTxs).reduce((acc, [, transactions]) => (acc += transactions.length), 0)
+      : 0
+    setTxsCount({ next, queued })
+  }, [nextTxs, queuedTxs])
+
+  return {
+    next: {
+      count: txsCount.next,
+      transactions: nextTxs ? Object.entries(nextTxs) : [],
+    },
+    queue: {
+      count: txsCount.queued,
+      transactions: queuedTxs ? Object.entries(queuedTxs) : [],
+    },
+  }
+}
