+++ src/routes/safe/components/GatewayTransactions/hooks/useTransactionType.ts
@@ -1,33 +1,20 @@
-import React, { useState, useEffect, FC, HTMLAttributes } from 'react'
+import { useEffect, useState } from 'react'
 import { useSelector } from 'react-redux'
 
-import CustomTxIcon from './assets/custom.svg'
-import IncomingTxIcon from './assets/incoming.svg'
-import OutgoingTxIcon from './assets/outgoing.svg'
-import SettingsTxIcon from './assets/settings.svg'
-
-import CustomIconText from 'src/components/CustomIconText'
-import { Custom, Transaction } from 'src/logic/safe/store/models/types/gateway'
+import { Transaction } from 'src/logic/safe/store/models/types/gateway'
 import { safeParamAddressFromStateSelector } from 'src/logic/safe/store/selectors'
-import { sameString } from 'src/utils/strings'
-import { sameAddress } from 'src/logic/wallets/ethAddresses'
-
-interface TxTypeProps {
-  tx: Transaction
-}
-
-const isCancelTransaction = ({ txInfo, safeAddress }: { txInfo: Custom; safeAddress: string }): boolean =>
-  sameAddress(txInfo.to, safeAddress) &&
-  sameString(txInfo.dataSize, '0') &&
-  sameString(txInfo.value, '0') &&
-  txInfo.methodName === null
+import CustomTxIcon from 'src/routes/safe/components/GatewayTransactions/assets/custom.svg'
+import IncomingTxIcon from 'src/routes/safe/components/GatewayTransactions/assets/incoming.svg'
+import OutgoingTxIcon from 'src/routes/safe/components/GatewayTransactions/assets/outgoing.svg'
+import SettingsTxIcon from 'src/routes/safe/components/GatewayTransactions/assets/settings.svg'
+import { isCancelTransaction } from 'src/routes/safe/components/GatewayTransactions/utils'
 
 type TxType = {
   icon: string
   text: string
 }
 
-export const useTxType = (tx: Transaction): TxType => {
+export const useTransactionType = (tx: Transaction): TxType => {
   const [type, setType] = useState<TxType>({ icon: CustomTxIcon, text: 'Custom transaction' })
   const safeAddress = useSelector(safeParamAddressFromStateSelector)
 
@@ -64,9 +51,4 @@ export const useTxType = (tx: Transaction): TxType => {
   }, [tx, safeAddress])
 
   return type
-}
-
-export const TxType: FC<TxTypeProps & HTMLAttributes<unknown>> = ({ tx }) => {
-  const { icon, text } = useTxType(tx)
-  return <CustomIconText iconUrl={icon} text={text} />
 }
